+++ CHANGELOG.md
@@ -43,6 +43,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- `canadensis_can`: Breaking change: CAN transport can successfully receive transfers that are larger than the
+  subscription's maximum transfer length  [#48](https://github.com/samcrow/canadensis/pull/48)
 - `canadensis_codegen_rust`: Deprecated modules are no longer deprecated for tests, which avoids irrelevant warnings
   when compiling tests